+++ core/src/main/kotlin/io/github/ackeecz/guardian/core/keystore/android/Utils.kt
@@ -1,0 +1,3 @@
+package io.github.ackeecz.guardian.core.keystore.android
+
+internal const val ANDROID_KEY_STORE_ID = "AndroidKeyStore"