+++ datastore/src/main/java/io/github/ackeecz/guardian/datastore/EncryptedDataStoreFactory.kt
@@ -8,17 +8,23 @@ import androidx.datastore.core.DataStoreFactory
 import androidx.datastore.core.Serializer
 import androidx.datastore.core.handlers.ReplaceFileCorruptionHandler
 import androidx.datastore.dataStoreFile
+import io.github.ackeecz.guardian.core.keystore.android.AndroidKeyStoreSemaphore
 import io.github.ackeecz.guardian.datastore.core.DataStoreCryptoParams
 import io.github.ackeecz.guardian.datastore.core.internal.EncryptingSerializer
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.sync.Semaphore
 import java.io.File
+import java.security.KeyStore
 
 /**
  * Creates an instance of encrypted [DataStore]. Never create more than one instance of
  * DataStore for a given file. Doing so can break all DataStore functionality. You should
  * consider managing your DataStore instance as a singleton.
+ *
+ * All Android [KeyStore] operations are synchronized using a provided [keyStoreSemaphore].
+ * More info about this topic can be found in [AndroidKeyStoreSemaphore] documentation.
  *
  * @param context Context of the application. It is safe to pass an Activity context, but best
  * practice is to pass Application one to be sure to avoid potential memory leaks.
@@ -31,6 +37,9 @@ import java.io.File
  * may be run more than once whether or not it already succeeded (potentially because another
  * migration failed or a write to disk failed.)
  * @param scope The scope in which IO operations and transform functions will execute.
+ * @param keyStoreSemaphore [Semaphore] used to synchronize Android [KeyStore] operations.
+ * It is recommended to use a default [AndroidKeyStoreSemaphore], if you really don't need
+ * to provide a custom [Semaphore].
  * @param produceFile Function which returns the file that the new DataStore will act on.
  * The function must return the same file every time. No two instances of DataStore
  * should act on the same file at the same time. File will be created if it doesn't exist. You can
@@ -38,6 +47,7 @@ import java.io.File
  *
  * @return a new encrypted DataStore instance with the provided configuration
  */
+@JvmOverloads
 @Suppress("LongParameterList")
 public fun <T> DataStoreFactory.createEncrypted(
     context: Context,
@@ -46,6 +56,7 @@ public fun <T> DataStoreFactory.createEncrypted(
     corruptionHandler: ReplaceFileCorruptionHandler<T>? = null,
     migrations: List<DataMigration<T>> = emptyList(),
     scope: CoroutineScope = CoroutineScope(Dispatchers.IO + SupervisorJob()),
+    keyStoreSemaphore: Semaphore = AndroidKeyStoreSemaphore,
     produceFile: () -> File,
 ): DataStore<T> {
     val dataStoreFile = produceFile()
@@ -54,6 +65,7 @@ public fun <T> DataStoreFactory.createEncrypted(
         cryptoParams = cryptoParams,
         delegate = serializer,
         scope = scope,
+        keyStoreSemaphore = keyStoreSemaphore,
         dataStoreFile = dataStoreFile,
     )
     return create(